requirements.txt
asgiref==3.6.0
backports.zoneinfo;python_version<"3.9"
Django==4.1.7
django-filter==22.1
djangorestframework==3.14.0
importlib-metadata==6.0.0
Markdown==3.4.1
mysqlclient==2.1.1
pytz==2022.7.1
sqlparse==0.4.3
zipp==3.13.0
